renamed 'heritage fetcher' to 'heritage object fetcher'
