+++ apps/researcher/src/lib/objects/fetcher-result.ts
