+++ apps/researcher/src/lib/objects/index.ts
@@ -1,0 +1,1 @@
+export * from './fetcher';
